refactor progress header
