+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "Successful Habits Flow 01",
-    "slug": "successful-habits-flow-01",
+    "name": "vision",
+    "slug": "vision",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
