+++ src/navigations/SucHab01Navigator/index.js
@@ -8,25 +8,49 @@ import PromptVideo from '../../screens/SucHab01Screens/PromptVideo'
 import PromptText from '../../screens/SucHab01Screens/PromptText'
 import PostActivity from '../../screens/SucHab01Screens/PostActivity'
 import End from '../../screens/SucHab01Screens/End'
+import Header from '../../components/Header'
+import BenefitProgressHeader from '../../components/BenefitProgressHeader'
 
 const Stack = createStackNavigator()
 const SucHab01Navigator = () => {
   return (
     <Stack.Navigator>
-      <Stack.Screen
-        name="SuccessfulHabits01"
-        component={SuccessfulHabits}
-        options={{ headerShown: false }}
-      />
+      <Stack.Screen name="SuccessfulHabits01" component={SuccessfulHabits} />
       <Stack.Screen
         name="PreActivity"
         component={PreActivity}
-        options={{ headerShown: false, presentation: 'modal' }}
+        options={({ navigation }) => ({
+          header: () => (
+            <Header navigation={navigation} title="Reflection activity" />
+          ),
+          presentation: 'modal'
+        })}
       />
       <Stack.Screen
         name="PromptAudio"
         component={PromptAudio}
-        options={{ headerShown: false }}
+        options={({ navigation }) => ({
+          header: () => (
+            <BenefitProgressHeader
+              current={2}
+              total={3}
+              navigation={navigation}
+            />
+          )
+        })}
+      />
+      <Stack.Screen
+        name="PromptAudio2"
+        component={PromptAudio}
+        options={({ navigation }) => ({
+          header: () => (
+            <BenefitProgressHeader
+              current={3}
+              total={3}
+              navigation={navigation}
+            />
+          )
+        })}
       />
       <Stack.Screen
         name="PromptVideo"
@@ -36,12 +60,41 @@ const SucHab01Navigator = () => {
       <Stack.Screen
         name="PromptText"
         component={PromptText}
-        options={{ headerShown: false }}
+        options={({ navigation }) => ({
+          header: () => (
+            <BenefitProgressHeader
+              current={2}
+              total={3}
+              navigation={navigation}
+            />
+          )
+        })}
       />
       <Stack.Screen
         name="Breathe"
         component={BreatheScreen}
-        options={{ headerShown: false }}
+        options={({ navigation }) => ({
+          header: () => (
+            <BenefitProgressHeader
+              current={1}
+              total={3}
+              navigation={navigation}
+            />
+          )
+        })}
+      />
+      <Stack.Screen
+        name="Breathe2"
+        component={BreatheScreen}
+        options={({ navigation }) => ({
+          header: () => (
+            <BenefitProgressHeader
+              current={2}
+              total={3}
+              navigation={navigation}
+            />
+          )
+        })}
       />
       <Stack.Screen
         name="PostActivity"
