--- src/screens/BenefitListFlow3/SelectionScreen.js
@@ -1,250 +1,0 @@
-import React, { useState } from 'react'
-import { SafeAreaView, View, Text } from 'react-native'
-import { TouchableOpacity } from 'react-native-gesture-handler'
-import { useNavigation } from '@react-navigation/native'
-import Svg, { Rect, Path, Circle } from 'react-native-svg'
-import FeatherIcons from 'react-native-vector-icons/Feather'
-
-const SelectionScreen = () => {
-  const navigation = useNavigation()
-  const [isAudio, setIsAudio] = useState(false)
-  const [stopAudio, setStopAudio] = useState(false)
-
-  const resetAudio = () => {
-    setIsAudio(false)
-    setStopAudio(false)
-  }
-
-  const renderAudio = () => {
-    return (
-      <View style={{ alignItems: 'center' }}>
-        <View
-          style={{
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'space-evenly',
-            width: '100%'
-          }}>
-          <TouchableOpacity onPress={resetAudio}>
-            <FeatherIcons name={'trash-2'} size={20} />
-          </TouchableOpacity>
-          {isAudio && !stopAudio && (
-            <TouchableOpacity onPress={() => setStopAudio(true)}>
-              <Svg
-                width={72}
-                height={72}
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg">
-                <Circle
-                  cx={36}
-                  cy={36}
-                  r={35.25}
-                  stroke="#193340"
-                  strokeOpacity={0.25}
-                  strokeWidth={1.5}
-                />
-                <Rect
-                  x={22.5}
-                  y={22.5}
-                  width={27}
-                  height={27}
-                  rx={4.5}
-                  fill="#FE3C31"
-                />
-              </Svg>
-            </TouchableOpacity>
-          )}
-          {isAudio && stopAudio && (
-            <TouchableOpacity onPress={() => setStopAudio(true)}>
-              <Svg
-                width={72}
-                height={72}
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg">
-                <Circle
-                  cx={36}
-                  cy={36}
-                  r={35.25}
-                  stroke="#193340"
-                  strokeOpacity={0.25}
-                  strokeWidth={1.5}
-                />
-                <Path
-                  d="M48.586 37.827c1.406-.812 1.406-2.842 0-3.654l-18.984-10.96c-1.407-.812-3.165.203-3.165 1.826v21.922c0 1.623 1.758 2.638 3.165 1.826l18.984-10.96z"
-                  fill="#193340"
-                />
-              </Svg>
-            </TouchableOpacity>
-          )}
-          {isAudio && stopAudio && (
-            <TouchableOpacity onPress={() => setIsAudio(false)}>
-              <FeatherIcons name={'rotate-ccw'} size={20} />
-            </TouchableOpacity>
-          )}
-        </View>
-      </View>
-    )
-  }
-
-  const renderText = () => {
-    return (
-      <View style={{ alignItems: 'center' }}>
-        <Text style={{ marginBottom: 16, color: '#737376' }}>
-          Talk out loud
-        </Text>
-        <View
-          style={{
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'space-evenly',
-            width: '100%'
-          }}>
-          <TouchableOpacity
-            onPress={() => navigation.navigate('PromptText')}
-            style={{
-              width: 30,
-              height: 30,
-              justifyContent: 'center',
-              alignItems: 'center'
-            }}>
-            <Svg
-              width={22}
-              height={14}
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg">
-              <Path
-                d="M20 1H2a1 1 0 00-1 1v10a1 1 0 001 1h18a1 1 0 001-1V2a1 1 0 00-1-1z"
-                stroke="#717173"
-                strokeWidth={1.5}
-                strokeLinejoin="round"
-              />
-              <Path
-                d="M6 5a1 1 0 100-2 1 1 0 000 2zM7 8a1 1 0 100-2 1 1 0 000 2zM4 8a1 1 0 100-2 1 1 0 000 2zM9 5a1 1 0 100-2 1 1 0 000 2zM10 8a1 1 0 100-2 1 1 0 000 2zM12 5a1 1 0 100-2 1 1 0 000 2zM13 8a1 1 0 100-2 1 1 0 000 2zM15 5a1 1 0 100-2 1 1 0 000 2zM16 8a1 1 0 100-2 1 1 0 000 2zM18 5a1 1 0 100-2 1 1 0 000 2z"
-                fill="#717173"
-              />
-              <Path
-                d="M7.5 10h7"
-                stroke="#717173"
-                strokeWidth={1.5}
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-            </Svg>
-          </TouchableOpacity>
-          <TouchableOpacity onPress={() => setIsAudio(true)}>
-            <Svg
-              width={72}
-              height={72}
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg">
-              <Rect width={72} height={72} rx={36} fill="#FE3C31" />
-              <Path
-                d="M36 22.25A3.75 3.75 0 0032.25 26v10a3.75 3.75 0 007.5 0V26A3.75 3.75 0 0036 22.25v0z"
-                stroke="#fff"
-                strokeWidth={2.5}
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-              <Path
-                d="M44.75 33.5V36a8.75 8.75 0 01-17.5 0v-2.5M36 44.75v5M31 49.75h10"
-                stroke="#fff"
-                strokeWidth={2.5}
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-            </Svg>
-          </TouchableOpacity>
-          <TouchableOpacity
-            style={{
-              width: 30,
-              height: 30,
-              justifyContent: 'center',
-              alignItems: 'center'
-            }}>
-            <Svg
-              width={24}
-              height={24}
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg">
-              <Path
-                d="M23 7l-7 5 7 5V7zM14 5H3a2 2 0 00-2 2v10a2 2 0 002 2h11a2 2 0 002-2V7a2 2 0 00-2-2z"
-                stroke="#717173"
-                strokeWidth={2}
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-            </Svg>
-          </TouchableOpacity>
-        </View>
-      </View>
-    )
-  }
-
-  return (
-    <SafeAreaView
-      style={{
-        backgroundColor: '#DCDCDD',
-        flex: 1,
-        alignItems: 'center',
-        justifyContent: 'space-between'
-      }}>
-      <Svg
-        width={375}
-        height={134}
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg">
-        <Path
-          d="M113.611 48.29c2.789 0 4.574-1.427 4.574-3.718v-.009c0-1.776-1.03-2.789-3.545-3.328l-1.303-.29c-1.552-.333-2.183-.914-2.183-1.802v-.008c0-1.13 1.038-1.785 2.44-1.785 1.453-.008 2.391.706 2.549 1.685l.016.108h1.802l-.009-.116c-.132-1.86-1.793-3.296-4.333-3.296s-4.333 1.428-4.341 3.495v.008c0 1.768 1.046 2.897 3.478 3.412l1.312.282c1.585.34 2.233.938 2.233 1.86v.008c0 1.112-1.088 1.876-2.615 1.876-1.602 0-2.756-.697-2.856-1.785l-.008-.091h-1.826l.008.108c.166 2.017 1.901 3.386 4.607 3.386zm6.728-.29h1.801v-5.288c0-1.452.822-2.357 2.142-2.357 1.32 0 1.942.739 1.942 2.233V48h1.793v-5.835c0-2.15-1.112-3.354-3.129-3.354-1.312 0-2.175.581-2.615 1.544h-.133v-4.906h-1.801V48zm13.842.174c2.639 0 4.258-1.768 4.258-4.673v-.017c0-2.905-1.627-4.673-4.258-4.673-2.64 0-4.267 1.776-4.267 4.673v.017c0 2.905 1.619 4.673 4.267 4.673zm0-1.502c-1.553 0-2.424-1.17-2.424-3.171v-.017c0-2 .871-3.17 2.424-3.17 1.544 0 2.424 1.17 2.424 3.17v.017c0 1.992-.88 3.17-2.424 3.17zM140.419 48h1.801v-5.387c0-1.303.938-2.15 2.316-2.15.34 0 .648.041.971.108v-1.66a3.952 3.952 0 00-.788-.083c-1.204 0-2.042.564-2.366 1.519h-.133v-1.362h-1.801V48zm10.505.058c.348 0 .68-.041.971-.091V46.53a5.94 5.94 0 01-.681.033c-.888 0-1.253-.398-1.253-1.37v-4.79h1.934v-1.419h-1.934v-2.274h-1.835v2.274h-1.411v1.42h1.411v5.221c0 1.735.814 2.432 2.798 2.432zM153.518 44h5.553v-1.652h-5.553V44zm11.077 4.06c.349 0 .681-.042.972-.092V46.53a5.94 5.94 0 01-.681.033c-.888 0-1.253-.398-1.253-1.37v-4.79h1.934v-1.419h-1.934v-2.274h-1.835v2.274h-1.411v1.42h1.411v5.221c0 1.735.814 2.432 2.797 2.432zm6.587.115c2.308 0 3.553-1.328 3.852-2.515l.016-.075-1.734.009-.034.066c-.215.465-.904 1.038-2.058 1.038-1.486 0-2.432-1.005-2.466-2.731h6.392v-.631c0-2.706-1.544-4.524-4.059-4.524s-4.142 1.884-4.142 4.698v.009c0 2.855 1.594 4.656 4.233 4.656zm-.083-7.885c1.22 0 2.125.78 2.266 2.39h-4.582c.158-1.552 1.088-2.39 2.316-2.39zM177.122 48h1.801v-5.387c0-1.303.938-2.15 2.316-2.15.34 0 .647.041.971.108v-1.66a3.963 3.963 0 00-.789-.083c-1.203 0-2.042.564-2.365 1.519h-.133v-1.362h-1.801V48zm6.769 0h1.801v-5.537c0-1.178.806-2.108 1.893-2.108 1.062 0 1.735.647 1.735 1.677V48h1.793v-5.703c0-1.087.755-1.942 1.901-1.942 1.178 0 1.743.614 1.743 1.86V48h1.793v-6.217c0-1.876-1.063-2.972-2.881-2.972-1.253 0-2.291.64-2.747 1.61h-.141c-.399-.97-1.27-1.61-2.499-1.61-1.178 0-2.058.573-2.457 1.56h-.133v-1.386h-1.801V48zm24.272.15c2.274 0 3.735-1.819 3.735-4.649v-.017c0-2.855-1.444-4.656-3.735-4.656-1.237 0-2.291.614-2.773 1.585h-.132V35.45h-1.802V48h1.802v-1.42h.132c.523.988 1.511 1.57 2.773 1.57zm-.506-1.536c-1.503 0-2.433-1.187-2.433-3.113v-.017c0-1.925.93-3.112 2.433-3.112 1.502 0 2.407 1.178 2.407 3.112v.017c0 1.934-.905 3.113-2.407 3.113zm10.006 1.56c2.308 0 3.553-1.328 3.852-2.515l.016-.075-1.734.009-.034.066c-.215.465-.904 1.038-2.058 1.038-1.486 0-2.432-1.005-2.465-2.731h6.391v-.631c0-2.706-1.544-4.524-4.059-4.524s-4.142 1.884-4.142 4.698v.009c0 2.855 1.594 4.656 4.233 4.656zm-.083-7.885c1.221 0 2.125.78 2.266 2.39h-4.582c.158-1.552 1.088-2.39 2.316-2.39zM223.603 48h1.801v-5.288c0-1.452.822-2.357 2.142-2.357 1.319 0 1.942.739 1.942 2.233V48h1.793v-5.835c0-2.15-1.112-3.354-3.129-3.354-1.312 0-2.175.581-2.615 1.544h-.133v-1.37h-1.801V48zm13.808.174c2.308 0 3.553-1.328 3.852-2.515l.016-.075-1.734.009-.034.066c-.215.465-.904 1.038-2.058 1.038-1.486 0-2.432-1.005-2.466-2.731h6.392v-.631c0-2.706-1.544-4.524-4.059-4.524s-4.142 1.884-4.142 4.698v.009c0 2.855 1.594 4.656 4.233 4.656zm-.083-7.885c1.22 0 2.125.78 2.266 2.39h-4.582c.158-1.552 1.088-2.39 2.316-2.39zM243.99 48h1.793v-7.595h1.984v-1.42h-2.001v-.863c0-.913.39-1.395 1.386-1.395.283 0 .523.017.689.042v-1.32a6.566 6.566 0 00-1.062-.083c-1.876 0-2.789.88-2.789 2.673v.946h-1.486v1.42h1.486V48zm6.57-10.625c.614 0 1.129-.506 1.129-1.12 0-.623-.515-1.13-1.129-1.13a1.13 1.13 0 00-1.129 1.13c0 .614.506 1.12 1.129 1.12zM249.655 48h1.793v-9.015h-1.793V48zm7.65.058c.348 0 .68-.041.971-.091V46.53a5.94 5.94 0 01-.681.033c-.888 0-1.253-.398-1.253-1.37v-4.79h1.934v-1.419h-1.934v-2.274h-1.835v2.274h-1.411v1.42h1.411v5.221c0 1.735.814 2.432 2.798 2.432zm6.047.116c2.141 0 3.735-1.137 3.735-2.805v-.017c0-1.303-.83-2.042-2.565-2.449l-1.427-.323c-1.013-.241-1.428-.59-1.428-1.154v-.017c0-.722.714-1.212 1.701-1.212 1.013 0 1.652.473 1.827 1.104v.017h1.718v-.008c-.158-1.47-1.478-2.499-3.536-2.499-2.042 0-3.503 1.129-3.503 2.698v.008c0 1.32.788 2.092 2.49 2.482l1.436.332c1.046.24 1.469.623 1.469 1.187v.017c0 .738-.772 1.245-1.884 1.245-1.071 0-1.727-.457-1.951-1.137l-.008-.009h-1.801v.009c.174 1.51 1.552 2.531 3.727 2.531z"
-          fill="#16161A"
-        />
-        <Path
-          d="M22.13 78h1.23v-8.757h3.179v-1.107H18.95v1.107h3.178V78zm7.94.13c1.736 0 2.788-.984 3.041-1.983l.014-.054h-1.19l-.027.061c-.198.445-.813.916-1.811.916-1.313 0-2.154-.888-2.188-2.413h5.305v-.465c0-2.2-1.217-3.691-3.22-3.691s-3.302 1.559-3.302 3.835v.007c0 2.31 1.272 3.787 3.377 3.787zm-.083-6.57c1.087 0 1.894.691 2.017 2.147H27.93c.13-1.401.964-2.146 2.057-2.146zm5.28 6.44h1.19V67.705h-1.19V78zm3.682 0h1.19V67.705h-1.19V78zm10.158.13c1.08 0 1.825-.444 2.187-1.23h.11V78h1.19v-7.37h-1.19v4.362c0 1.292-.69 2.085-1.99 2.085-1.175 0-1.654-.636-1.654-1.962v-4.484h-1.19v4.771c0 1.743.862 2.728 2.537 2.728zm8.377 0c1.675 0 2.953-.91 2.953-2.222v-.014c0-1.052-.67-1.654-2.057-1.989l-1.135-.273c-.868-.212-1.237-.526-1.237-1.026v-.013c0-.65.642-1.1 1.517-1.1.889 0 1.463.403 1.62.984h1.17c-.165-1.204-1.238-1.976-2.783-1.976-1.565 0-2.741.923-2.741 2.146v.007c0 1.06.622 1.661 2.003 1.99l1.141.273c.91.219 1.285.567 1.285 1.066v.014c0 .67-.704 1.142-1.695 1.142-.943 0-1.531-.404-1.73-1.019H54.58c.137 1.217 1.265 2.01 2.905 2.01zm11.04 0c.991 0 1.764-.43 2.228-1.217h.11V78h1.19v-5.045c0-1.531-1.006-2.454-2.803-2.454-1.573 0-2.694.78-2.885 1.934l-.007.042h1.19l.006-.02c.192-.575.773-.903 1.654-.903 1.101 0 1.655.492 1.655 1.401v.67l-2.112.13c-1.716.102-2.687.861-2.687 2.174v.013c0 1.34 1.06 2.188 2.46 2.188zm-1.244-2.215v-.014c0-.731.492-1.128 1.613-1.196l1.969-.123v.67c0 1.053-.882 1.846-2.092 1.846-.855 0-1.49-.438-1.49-1.183zM74.433 78h1.19V67.705h-1.19V78zm3.681 0h1.19V67.705h-1.19V78zm10.309.055c.232 0 .458-.028.69-.069v-1.011c-.219.02-.335.027-.547.027-.765 0-1.066-.349-1.066-1.217v-4.17h1.613v-.984H87.5v-1.907h-1.23v1.907h-1.163v.984h1.162v4.47c0 1.41.636 1.97 2.154 1.97zM91.099 78h1.19v-4.361c0-1.292.745-2.085 1.92-2.085 1.176 0 1.723.635 1.723 1.962V78h1.19v-4.772c0-1.75-.923-2.727-2.578-2.727-1.087 0-1.777.458-2.146 1.237h-.11v-4.033h-1.19V78zm11.324.13c1.736 0 2.789-.984 3.042-1.983l.013-.054h-1.189l-.028.061c-.198.445-.813.916-1.811.916-1.313 0-2.153-.888-2.188-2.413h5.305v-.465c0-2.2-1.217-3.691-3.22-3.691s-3.301 1.559-3.301 3.835v.007c0 2.31 1.271 3.787 3.377 3.787zm-.082-6.57c1.086 0 1.893.691 2.016 2.147h-4.074c.13-1.401.964-2.146 2.058-2.146zm11.819 6.57c1.674 0 2.953-.91 2.953-2.222v-.014c0-1.052-.67-1.654-2.058-1.989l-1.135-.273c-.868-.212-1.237-.526-1.237-1.026v-.013c0-.65.643-1.1 1.518-1.1.888 0 1.462.403 1.62.984h1.169c-.164-1.204-1.238-1.976-2.783-1.976-1.565 0-2.741.923-2.741 2.146v.007c0 1.06.622 1.661 2.003 1.99l1.142.273c.909.219 1.285.567 1.285 1.066v.014c0 .67-.704 1.142-1.695 1.142-.944 0-1.532-.404-1.73-1.019h-1.217c.137 1.217 1.265 2.01 2.906 2.01zm5.007-.13h1.189v-4.361c0-1.292.745-2.085 1.921-2.085 1.176 0 1.723.635 1.723 1.962V78h1.189v-4.772c0-1.75-.923-2.727-2.577-2.727-1.087 0-1.777.458-2.146 1.237h-.11v-4.033h-1.189V78zm11.344.13c2.098 0 3.397-1.45 3.397-3.808v-.013c0-2.366-1.299-3.808-3.397-3.808-2.099 0-3.398 1.442-3.398 3.808v.013c0 2.359 1.299 3.808 3.398 3.808zm0-1.053c-1.395 0-2.181-1.018-2.181-2.755v-.013c0-1.744.786-2.755 2.181-2.755 1.394 0 2.181 1.011 2.181 2.755v.013c0 1.737-.787 2.755-2.181 2.755zm5.383.923h1.189v-4.566c0-1.08.807-1.798 1.942-1.798.26 0 .485.027.731.068V70.55a3.99 3.99 0 00-.594-.048c-.998 0-1.689.451-1.969 1.224h-.11V70.63h-1.189V78zm8.5.055c.233 0 .458-.028.691-.069v-1.011c-.219.02-.335.027-.547.027-.766 0-1.066-.349-1.066-1.217v-4.17h1.613v-.984h-1.613v-1.907h-1.231v1.907h-1.162v.984h1.162v4.47c0 1.41.636 1.97 2.153 1.97zm2.232-3.473h4.594v-1.107h-4.594v1.107zm9.191 3.473c.232 0 .458-.028.69-.069v-1.011c-.218.02-.335.027-.546.027-.766 0-1.067-.349-1.067-1.217v-4.17h1.613v-.984h-1.613v-1.907h-1.23v1.907h-1.162v.984h1.162v4.47c0 1.41.635 1.97 2.153 1.97zm5.506.075c1.736 0 2.789-.984 3.042-1.983l.014-.054h-1.19l-.027.061c-.198.445-.813.916-1.811.916-1.313 0-2.154-.888-2.188-2.413h5.305v-.465c0-2.2-1.217-3.691-3.22-3.691s-3.302 1.559-3.302 3.835v.007c0 2.31 1.272 3.787 3.377 3.787zm-.082-6.57c1.087 0 1.894.691 2.017 2.147h-4.074c.129-1.401.963-2.146 2.057-2.146zm5.212 6.44h1.19v-4.566c0-1.08.806-1.798 1.941-1.798.26 0 .486.027.732.068V70.55a4.005 4.005 0 00-.595-.048c-.998 0-1.688.451-1.969 1.224h-.109V70.63h-1.19V78zm5.609 0h1.19v-4.566c0-1.04.731-1.88 1.695-1.88.93 0 1.531.567 1.531 1.442V78h1.19v-4.737c0-.937.676-1.71 1.702-1.71 1.039 0 1.538.54 1.538 1.628V78h1.189v-5.093c0-1.545-.841-2.406-2.344-2.406-1.019 0-1.86.513-2.256 1.292h-.11c-.341-.766-1.039-1.292-2.037-1.292-.964 0-1.661.458-1.989 1.251h-.109v-1.121h-1.19V78zm20.063.13c1.867 0 3.09-1.51 3.09-3.808v-.013c0-2.311-1.217-3.808-3.09-3.808-1.011 0-1.893.506-2.297 1.292h-.109v-4.088h-1.189V78h1.189v-1.176h.109c.452.82 1.272 1.306 2.297 1.306zm-.273-1.053c-1.34 0-2.16-1.053-2.16-2.755v-.013c0-1.703.82-2.755 2.16-2.755 1.347 0 2.146 1.039 2.146 2.755v.013c0 1.716-.799 2.755-2.146 2.755zm8.357 1.053c1.736 0 2.789-.984 3.042-1.983l.013-.054h-1.189l-.027.061c-.199.445-.814.916-1.812.916-1.312 0-2.153-.888-2.187-2.413h5.304v-.465c0-2.2-1.217-3.691-3.219-3.691-2.003 0-3.302 1.559-3.302 3.835v.007c0 2.31 1.271 3.787 3.377 3.787zm-.082-6.57c1.087 0 1.893.691 2.016 2.147h-4.074c.13-1.401.964-2.146 2.058-2.146zm5.212 6.44h1.189v-4.361c0-1.292.746-2.085 1.921-2.085 1.176 0 1.723.635 1.723 1.962V78h1.189v-4.772c0-1.75-.922-2.727-2.577-2.727-1.087 0-1.777.458-2.146 1.237h-.11v-1.107h-1.189V78zm11.324.13c1.736 0 2.789-.984 3.042-1.983l.013-.054h-1.189l-.028.061c-.198.445-.813.916-1.811.916-1.313 0-2.153-.888-2.188-2.413h5.305v-.465c0-2.2-1.217-3.691-3.22-3.691s-3.301 1.559-3.301 3.835v.007c0 2.31 1.271 3.787 3.377 3.787zm-.082-6.57c1.086 0 1.893.691 2.016 2.147h-4.074c.13-1.401.964-2.146 2.058-2.146zm5.704 6.44h1.189v-6.385h1.675v-.984h-1.675v-.786c0-.807.342-1.224 1.19-1.224.212 0 .41.007.554.034v-.95a4.133 4.133 0 00-.793-.068c-1.402 0-2.14.704-2.14 2.174v.82h-1.224v.984h1.224V78zm5.349-8.791c.451 0 .82-.37.82-.82a.823.823 0 00-.82-.82c-.451 0-.82.368-.82.82 0 .45.369.82.82.82zM227.032 78h1.19v-7.37h-1.19V78zm6.19.055c.233 0 .458-.028.691-.069v-1.011c-.219.02-.335.027-.547.027-.766 0-1.067-.349-1.067-1.217v-4.17h1.614v-.984h-1.614v-1.907h-1.23v1.907h-1.162v.984h1.162v4.47c0 1.41.636 1.97 2.153 1.97zm5.001.075c1.674 0 2.953-.91 2.953-2.222v-.014c0-1.052-.67-1.654-2.058-1.989l-1.135-.273c-.868-.212-1.237-.526-1.237-1.026v-.013c0-.65.643-1.1 1.518-1.1.888 0 1.462.403 1.62.984h1.169c-.164-1.204-1.238-1.976-2.783-1.976-1.565 0-2.741.923-2.741 2.146v.007c0 1.06.622 1.661 2.003 1.99l1.142.273c.909.219 1.285.567 1.285 1.066v.014c0 .67-.704 1.142-1.695 1.142-.944 0-1.532-.404-1.73-1.019h-1.217c.137 1.217 1.265 2.01 2.906 2.01zm11.634-.075c.232 0 .458-.028.69-.069v-1.011c-.218.02-.335.027-.546.027-.766 0-1.067-.349-1.067-1.217v-4.17h1.613v-.984h-1.613v-1.907h-1.23v1.907h-1.162v.984h1.162v4.47c0 1.41.635 1.97 2.153 1.97zm2.676-.055h1.19v-4.361c0-1.292.745-2.085 1.921-2.085 1.175 0 1.722.635 1.722 1.962V78h1.19v-4.772c0-1.75-.923-2.727-2.578-2.727-1.086 0-1.777.458-2.146 1.237h-.109v-4.033h-1.19V78zm10.339.13c.992 0 1.764-.43 2.229-1.217h.109V78h1.19v-5.045c0-1.531-1.005-2.454-2.803-2.454-1.572 0-2.693.78-2.885 1.934l-.007.042h1.19l.007-.02c.191-.575.772-.903 1.654-.903 1.101 0 1.654.492 1.654 1.401v.67l-2.112.13c-1.716.102-2.687.861-2.687 2.174v.013c0 1.34 1.06 2.188 2.461 2.188zm-1.244-2.215v-.014c0-.731.492-1.128 1.614-1.196l1.968-.123v.67c0 1.053-.882 1.846-2.092 1.846-.854 0-1.49-.438-1.49-1.183zm9.704 2.14c.232 0 .458-.028.69-.069v-1.011c-.219.02-.335.027-.547.027-.765 0-1.066-.349-1.066-1.217v-4.17h1.613v-.984h-1.613v-1.907h-1.231v1.907h-1.162v.984h1.162v4.47c0 1.41.636 1.97 2.154 1.97zm7.314 2.529c1.306 0 1.914-.478 2.522-2.133l2.878-7.82h-1.251l-2.016 6.063h-.11l-2.023-6.063h-1.272l2.728 7.376-.137.437c-.266.848-.677 1.156-1.353 1.156a2.82 2.82 0 01-.492-.035v.978c.164.027.369.041.526.041zm9.867-2.454c2.099 0 3.398-1.45 3.398-3.808v-.013c0-2.366-1.299-3.808-3.398-3.808-2.098 0-3.397 1.442-3.397 3.808v.013c0 2.359 1.299 3.808 3.397 3.808zm0-1.053c-1.394 0-2.18-1.018-2.18-2.755v-.013c0-1.744.786-2.755 2.18-2.755 1.395 0 2.181 1.011 2.181 2.755v.013c0 1.737-.786 2.755-2.181 2.755zm7.851 1.053c1.08 0 1.826-.444 2.188-1.23h.109V78h1.19v-7.37h-1.19v4.362c0 1.292-.69 2.085-1.989 2.085-1.176 0-1.654-.636-1.654-1.962v-4.484h-1.19v4.771c0 1.743.862 2.728 2.536 2.728zm10.555-8.921c.451 0 .82-.37.82-.82a.823.823 0 00-.82-.82.823.823 0 00-.821.82c0 .45.37.82.821.82zM306.317 78h1.19v-7.37h-1.19V78zm3.572 0h1.189v-4.566c0-1.04.732-1.88 1.696-1.88.929 0 1.531.567 1.531 1.442V78h1.189v-4.737c0-.937.677-1.71 1.702-1.71 1.04 0 1.539.54 1.539 1.628V78h1.189v-5.093c0-1.545-.841-2.406-2.345-2.406-1.018 0-1.859.513-2.256 1.292h-.109c-.342-.766-1.039-1.292-2.037-1.292-.964 0-1.661.458-1.989 1.251h-.11v-1.121h-1.189V78zm14.345.13c.991 0 1.764-.43 2.228-1.217h.11V78h1.189v-5.045c0-1.531-1.005-2.454-2.802-2.454-1.573 0-2.694.78-2.885 1.934l-.007.042h1.189l.007-.02c.192-.575.773-.903 1.655-.903 1.1 0 1.654.492 1.654 1.401v.67l-2.112.13c-1.716.102-2.687.861-2.687 2.174v.013c0 1.34 1.06 2.188 2.461 2.188zm-1.244-2.215v-.014c0-.731.492-1.128 1.613-1.196l1.969-.123v.67c0 1.053-.882 1.846-2.092 1.846-.854 0-1.49-.438-1.49-1.183zm10.12 4.683c2.01 0 3.282-1.053 3.282-2.7V70.63h-1.19v1.217h-.082c-.451-.834-1.258-1.347-2.297-1.347-1.928 0-3.117 1.497-3.117 3.561v.014c0 2.065 1.183 3.541 3.09 3.541 1.012 0 1.846-.458 2.31-1.271h.11v1.49c0 1.087-.786 1.709-2.106 1.709-1.059 0-1.716-.397-1.845-.957l-.007-.007h-1.231l-.013.007c.184 1.21 1.319 2.01 3.096 2.01zm-.054-4.034c-1.368 0-2.133-1.025-2.133-2.488v-.013c0-1.463.765-2.51 2.133-2.51 1.36 0 2.187 1.047 2.187 2.51v.013c0 1.463-.82 2.489-2.187 2.489zm6.319-7.355c.451 0 .821-.37.821-.82 0-.451-.37-.82-.821-.82-.451 0-.82.368-.82.82 0 .45.369.82.82.82zM338.774 78h1.189v-7.37h-1.189V78zm3.571 0h1.19v-4.361c0-1.292.745-2.085 1.921-2.085 1.175 0 1.722.635 1.722 1.962V78h1.19v-4.772c0-1.75-.923-2.727-2.577-2.727-1.087 0-1.778.458-2.147 1.237h-.109v-1.107h-1.19V78zm11.324.13c1.736 0 2.789-.984 3.042-1.983l.014-.054h-1.19l-.027.061c-.198.445-.814.916-1.812.916-1.312 0-2.153-.888-2.187-2.413h5.304v-.465c0-2.2-1.216-3.691-3.219-3.691s-3.302 1.559-3.302 3.835v.007c0 2.31 1.271 3.787 3.377 3.787zm-.082-6.57c1.087 0 1.893.691 2.017 2.147h-4.075c.13-1.401.964-2.146 2.058-2.146zM20.157 99.599c2.01 0 3.281-1.053 3.281-2.7V89.63h-1.19v1.217h-.081c-.452-.834-1.258-1.347-2.297-1.347-1.928 0-3.117 1.497-3.117 3.561v.014c0 2.065 1.182 3.541 3.09 3.541 1.011 0 1.845-.458 2.31-1.271h.11v1.49c0 1.087-.787 1.709-2.106 1.709-1.06 0-1.716-.397-1.846-.957l-.007-.007h-1.23l-.014.007c.185 1.21 1.32 2.01 3.097 2.01zm-.055-4.034c-1.367 0-2.133-1.025-2.133-2.488v-.013c0-1.463.766-2.51 2.133-2.51 1.36 0 2.188 1.047 2.188 2.51v.013c0 1.463-.82 2.489-2.188 2.489zm7.715 1.566c.99 0 1.763-.43 2.228-1.217h.11V97h1.189v-5.045c0-1.531-1.005-2.454-2.803-2.454-1.572 0-2.693.78-2.885 1.934l-.006.042h1.189l.007-.02c.191-.575.772-.903 1.654-.903 1.1 0 1.654.492 1.654 1.401v.67l-2.112.13c-1.716.102-2.687.861-2.687 2.174v.013c0 1.34 1.06 2.188 2.462 2.188zm-1.245-2.215v-.014c0-.731.492-1.128 1.614-1.196l1.968-.123v.67c0 1.053-.881 1.846-2.091 1.846-.855 0-1.49-.438-1.49-1.183zm7.715-6.706c.45 0 .82-.37.82-.82 0-.451-.37-.82-.82-.82-.452 0-.82.368-.82.82 0 .45.368.82.82.82zM33.685 97h1.19v-7.37h-1.19V97zm3.572 0h1.19v-4.361c0-1.292.744-2.085 1.92-2.085s1.723.635 1.723 1.962V97h1.19v-4.772c0-1.75-.924-2.727-2.578-2.727-1.087 0-1.777.458-2.146 1.237h-.11v-1.107h-1.19V97zm8.944-8.791c.452 0 .82-.37.82-.82 0-.451-.368-.82-.82-.82-.45 0-.82.368-.82.82 0 .45.37.82.82.82zM45.6 97h1.19v-7.37H45.6V97zm3.571 0h1.19v-4.361c0-1.292.745-2.085 1.92-2.085 1.177 0 1.723.635 1.723 1.962V97h1.19v-4.772c0-1.75-.923-2.727-2.577-2.727-1.087 0-1.778.458-2.147 1.237h-.11v-1.107h-1.189V97zm11.351 2.598c2.01 0 3.282-1.053 3.282-2.7V89.63h-1.19v1.217h-.082c-.451-.834-1.258-1.347-2.297-1.347-1.928 0-3.117 1.497-3.117 3.561v.014c0 2.065 1.183 3.541 3.09 3.541 1.012 0 1.846-.458 2.31-1.271h.11v1.49c0 1.087-.786 1.709-2.106 1.709-1.06 0-1.715-.397-1.845-.957l-.007-.007h-1.23l-.014.007c.184 1.21 1.319 2.01 3.096 2.01zm-.054-4.034c-1.368 0-2.133-1.025-2.133-2.488v-.013c0-1.463.765-2.51 2.133-2.51 1.36 0 2.187 1.047 2.187 2.51v.013c0 1.463-.82 2.489-2.187 2.489zM70.728 97h1.19v-6.385h1.674v-.984h-1.674v-.786c0-.807.341-1.224 1.189-1.224.212 0 .41.007.554.034v-.95a4.124 4.124 0 00-.793-.068c-1.402 0-2.14.704-2.14 2.174v.82h-1.224v.984h1.224V97zm4.72 0h1.19v-4.566c0-1.08.806-1.798 1.941-1.798.26 0 .485.027.732.068V89.55a3.995 3.995 0 00-.595-.048c-.998 0-1.689.451-1.969 1.224h-.11V89.63h-1.189V97zm8.296.13c2.098 0 3.397-1.45 3.397-3.808v-.013c0-2.366-1.299-3.808-3.398-3.808-2.098 0-3.397 1.442-3.397 3.808v.013c0 2.359 1.299 3.808 3.397 3.808zm0-1.053c-1.395 0-2.181-1.018-2.181-2.755v-.013c0-1.744.786-2.755 2.18-2.755 1.395 0 2.181 1.011 2.181 2.755v.013c0 1.737-.786 2.755-2.18 2.755zm5.383.923h1.19v-4.566c0-1.04.73-1.88 1.694-1.88.93 0 1.532.567 1.532 1.442V97h1.19v-4.737c0-.937.676-1.71 1.701-1.71 1.04 0 1.538.54 1.538 1.628V97h1.19v-5.093c0-1.545-.841-2.406-2.345-2.406-1.018 0-1.86.513-2.256 1.292h-.11c-.341-.766-1.038-1.292-2.036-1.292-.964 0-1.661.458-1.99 1.251h-.109v-1.121h-1.19V97zm19.413.13c1.771 0 2.762-.95 3.063-2.283l.014-.075-1.176.006-.014.041c-.273.82-.902 1.258-1.893 1.258-1.313 0-2.161-1.087-2.161-2.782v-.014c0-1.66.834-2.727 2.161-2.727 1.059 0 1.715.588 1.9 1.312l.007.02h1.182l-.006-.04c-.219-1.313-1.292-2.345-3.083-2.345-2.065 0-3.377 1.49-3.377 3.78v.014c0 2.338 1.319 3.835 3.383 3.835zm7.988 0c2.099 0 3.398-1.45 3.398-3.808v-.013c0-2.366-1.299-3.808-3.398-3.808s-3.397 1.442-3.397 3.808v.013c0 2.359 1.298 3.808 3.397 3.808zm0-1.053c-1.394 0-2.181-1.018-2.181-2.755v-.013c0-1.744.787-2.755 2.181-2.755 1.395 0 2.181 1.011 2.181 2.755v.013c0 1.737-.786 2.755-2.181 2.755zm5.383.923h1.19v-4.566c0-1.04.731-1.88 1.695-1.88.93 0 1.531.567 1.531 1.442V97h1.19v-4.737c0-.937.676-1.71 1.702-1.71 1.039 0 1.538.54 1.538 1.628V97h1.189v-5.093c0-1.545-.84-2.406-2.344-2.406-1.019 0-1.86.513-2.256 1.292h-.11c-.341-.766-1.039-1.292-2.037-1.292-.964 0-1.661.458-1.989 1.251h-.109v-1.121h-1.19V97zm12.322 2.46h1.189v-3.622h.11c.403.786 1.285 1.292 2.297 1.292 1.873 0 3.089-1.497 3.089-3.808v-.013c0-2.297-1.223-3.808-3.089-3.808-1.026 0-1.846.485-2.297 1.306h-.11V89.63h-1.189v9.83zm3.322-3.383c-1.34 0-2.16-1.053-2.16-2.755v-.013c0-1.703.82-2.755 2.16-2.755 1.347 0 2.147 1.039 2.147 2.755v.013c0 1.716-.8 2.755-2.147 2.755zm5.431.923h1.19V86.705h-1.19V97zm6.621.13c1.736 0 2.789-.984 3.042-1.983l.013-.054h-1.189l-.027.061c-.199.445-.814.916-1.812.916-1.313 0-2.153-.888-2.188-2.413h5.305v-.465c0-2.2-1.217-3.691-3.22-3.691-2.002 0-3.301 1.559-3.301 3.835v.007c0 2.31 1.271 3.787 3.377 3.787zm-.082-6.57c1.087 0 1.893.691 2.016 2.147h-4.074c.13-1.401.964-2.146 2.058-2.146zm7.83 6.495c.232 0 .458-.028.69-.069v-1.011c-.218.02-.334.027-.546.027-.766 0-1.067-.349-1.067-1.217v-4.17h1.613v-.984h-1.613v-1.907h-1.23v1.907h-1.162v.984h1.162v4.47c0 1.41.635 1.97 2.153 1.97zm3.237-8.846c.451 0 .82-.37.82-.82a.823.823 0 00-.82-.82.823.823 0 00-.821.82c0 .45.37.82.821.82zM159.99 97h1.19v-7.37h-1.19V97zm3.572 0h1.189v-4.361c0-1.292.745-2.085 1.921-2.085 1.176 0 1.723.635 1.723 1.962V97h1.189v-4.772c0-1.75-.923-2.727-2.577-2.727-1.087 0-1.777.458-2.146 1.237h-.11v-1.107h-1.189V97zm11.351 2.598c2.01 0 3.281-1.053 3.281-2.7V89.63h-1.189v1.217h-.082c-.452-.834-1.258-1.347-2.297-1.347-1.928 0-3.118 1.497-3.118 3.561v.014c0 2.065 1.183 3.541 3.09 3.541 1.012 0 1.846-.458 2.311-1.271h.109v1.49c0 1.087-.786 1.709-2.105 1.709-1.06 0-1.716-.397-1.846-.957l-.007-.007h-1.23l-.014.007c.185 1.21 1.319 2.01 3.097 2.01zm-.055-4.034c-1.367 0-2.133-1.025-2.133-2.488v-.013c0-1.463.766-2.51 2.133-2.51 1.36 0 2.188 1.047 2.188 2.51v.013c0 1.463-.821 2.489-2.188 2.489zm10.397 4.02c1.306 0 1.914-.478 2.523-2.133l2.878-7.82h-1.251l-2.017 6.063h-.109l-2.024-6.063h-1.271l2.727 7.376-.136.437c-.267.848-.677 1.156-1.354 1.156a2.82 2.82 0 01-.492-.035v.978c.164.027.369.041.526.041zm9.868-2.454c2.098 0 3.397-1.45 3.397-3.808v-.013c0-2.366-1.299-3.808-3.397-3.808-2.099 0-3.398 1.442-3.398 3.808v.013c0 2.359 1.299 3.808 3.398 3.808zm0-1.053c-1.395 0-2.181-1.018-2.181-2.755v-.013c0-1.744.786-2.755 2.181-2.755 1.394 0 2.18 1.011 2.18 2.755v.013c0 1.737-.786 2.755-2.18 2.755zm7.851 1.053c1.08 0 1.825-.444 2.187-1.23h.11V97h1.189v-7.37h-1.189v4.362c0 1.292-.691 2.085-1.99 2.085-1.175 0-1.654-.636-1.654-1.962v-4.484h-1.189v4.771c0 1.743.861 2.728 2.536 2.728zm5.848-.13h1.189v-4.566c0-1.08.807-1.798 1.942-1.798.259 0 .485.027.731.068V89.55a3.992 3.992 0 00-.595-.048c-.998 0-1.688.451-1.968 1.224h-.11V89.63h-1.189V97zm11.744 0h1.162l2.269-7.786h.082l2.27 7.786h1.162l2.686-9.864h-1.285l-1.962 8.018h-.082l-2.208-8.018h-1.244l-2.208 8.018h-.082l-1.962-8.018h-1.285L220.566 97zm13.818.13c1.737 0 2.789-.984 3.042-1.983l.014-.054h-1.189l-.028.061c-.198.445-.813.916-1.811.916-1.313 0-2.154-.888-2.188-2.413h5.305v-.465c0-2.2-1.217-3.691-3.22-3.691s-3.302 1.559-3.302 3.835v.007c0 2.31 1.272 3.787 3.377 3.787zm-.082-6.57c1.087 0 1.894.691 2.017 2.147h-4.074c.13-1.401.964-2.146 2.057-2.146zm8.22 6.57c1.737 0 2.789-.984 3.042-1.983l.014-.054h-1.189l-.028.061c-.198.445-.813.916-1.811.916-1.313 0-2.154-.888-2.188-2.413h5.305v-.465c0-2.2-1.217-3.691-3.22-3.691s-3.302 1.559-3.302 3.835v.007c0 2.31 1.272 3.787 3.377 3.787zm-.082-6.57c1.087 0 1.894.691 2.017 2.147h-4.074c.13-1.401.964-2.146 2.057-2.146zm5.281 6.44h1.189v-2.796l.691-.67L252.363 97h1.497l-3.425-4.272 3.213-3.097h-1.443l-3.185 3.22h-.11v-6.146h-1.189V97zm7.741 0h1.19V86.705h-1.19V97zm4.242 2.584c1.306 0 1.914-.478 2.523-2.133l2.877-7.82h-1.251l-2.016 6.063h-.11l-2.023-6.063h-1.271l2.727 7.376-.137.437c-.266.848-.676 1.156-1.353 1.156a2.82 2.82 0 01-.492-.035v.978c.164.027.369.041.526.041zm15.312-2.352c2.099 0 3.658-1.168 3.986-2.973v-.014h-1.245l-.013.014c-.308 1.135-1.354 1.839-2.728 1.839-1.9 0-3.09-1.552-3.09-4.02v-.014c0-2.467 1.19-4.026 3.083-4.026 1.361 0 2.414.8 2.742 2.065l.006.013h1.238l.007-.013c-.294-1.894-1.914-3.2-3.993-3.2-2.679 0-4.347 1.983-4.347 5.161v.014c0 3.172 1.675 5.154 4.354 5.154zm9.068-.102c2.099 0 3.397-1.45 3.397-3.808v-.013c0-2.366-1.298-3.808-3.397-3.808-2.099 0-3.397 1.442-3.397 3.808v.013c0 2.359 1.298 3.808 3.397 3.808zm0-1.053c-1.395 0-2.181-1.018-2.181-2.755v-.013c0-1.744.786-2.755 2.181-2.755s2.181 1.011 2.181 2.755v.013c0 1.737-.786 2.755-2.181 2.755zm5.383.923h1.19v-4.566c0-1.04.731-1.88 1.695-1.88.93 0 1.531.567 1.531 1.442V97h1.19v-4.737c0-.937.676-1.71 1.702-1.71 1.039 0 1.538.54 1.538 1.628V97h1.189v-5.093c0-1.545-.84-2.406-2.344-2.406-1.019 0-1.86.513-2.256 1.292h-.11c-.341-.766-1.039-1.292-2.037-1.292-.964 0-1.661.458-1.989 1.251h-.109v-1.121h-1.19V97zm12.322 0h1.189v-4.566c0-1.04.732-1.88 1.696-1.88.929 0 1.531.567 1.531 1.442V97h1.189v-4.737c0-.937.677-1.71 1.702-1.71 1.039 0 1.539.54 1.539 1.628V97h1.189v-5.093c0-1.545-.841-2.406-2.345-2.406-1.018 0-1.859.513-2.256 1.292h-.109c-.342-.766-1.039-1.292-2.037-1.292-.964 0-1.661.458-1.989 1.251h-.11v-1.121h-1.189V97zm12.95-8.791c.452 0 .821-.37.821-.82a.823.823 0 00-.821-.82c-.451 0-.82.368-.82.82 0 .45.369.82.82.82zM314.138 97h1.189v-7.37h-1.189V97zm6.19.055c.232 0 .458-.028.69-.069v-1.011c-.219.02-.335.027-.547.027-.765 0-1.066-.349-1.066-1.217v-4.17h1.613v-.984h-1.613v-1.907h-1.231v1.907h-1.162v.984h1.162v4.47c0 1.41.636 1.97 2.154 1.97zm2.607-.055h1.19v-4.566c0-1.04.731-1.88 1.695-1.88.93 0 1.531.567 1.531 1.442V97h1.19v-4.737c0-.937.677-1.71 1.702-1.71 1.039 0 1.538.54 1.538 1.628V97h1.19v-5.093c0-1.545-.841-2.406-2.345-2.406-1.019 0-1.86.513-2.256 1.292h-.109c-.342-.766-1.039-1.292-2.038-1.292-.963 0-1.661.458-1.989 1.251h-.109v-1.121h-1.19V97zm15.33.13c1.736 0 2.789-.984 3.042-1.983l.013-.054h-1.189l-.027.061c-.199.445-.814.916-1.812.916-1.312 0-2.153-.888-2.187-2.413h5.304v-.465c0-2.2-1.216-3.691-3.219-3.691s-3.302 1.559-3.302 3.835v.007c0 2.31 1.271 3.787 3.377 3.787zm-.082-6.57c1.087 0 1.893.691 2.016 2.147h-4.074c.13-1.401.964-2.146 2.058-2.146zm5.212 6.44h1.19v-4.361c0-1.292.745-2.085 1.92-2.085 1.176 0 1.723.635 1.723 1.962V97h1.19v-4.772c0-1.75-.923-2.727-2.578-2.727-1.087 0-1.777.458-2.146 1.237h-.109v-1.107h-1.19V97zm10.934.055c.232 0 .458-.028.69-.069v-1.011c-.218.02-.335.027-.546.027-.766 0-1.067-.349-1.067-1.217v-4.17h1.613v-.984h-1.613v-1.907h-1.23v1.907h-1.162v.984h1.162v4.47c0 1.41.635 1.97 2.153 1.97zm3.579.013a.89.89 0 00.888-.888.886.886 0 00-.888-.889.891.891 0 00-.889.889c0 .485.403.888.889.888z"
-          fill="#4D4D4D"
-        />
-      </Svg>
-      {isAudio ? renderAudio() : renderText()}
-      <View style={{ flexDirection: 'row', padding: 16 }}>
-        <TouchableOpacity
-          onPress={navigation.goBack}
-          style={{
-            marginRight: 4,
-            borderRadius: 100,
-            width: 168,
-            height: 52,
-            justifyContent: 'center',
-            backgroundColor: 'rgba(25, 51, 64, 0.08)'
-          }}>
-          <Text
-            style={{
-              fontSize: 17,
-              fontWeight: '500',
-              textAlign: 'center'
-            }}>
-            Back
-          </Text>
-        </TouchableOpacity>
-        <TouchableOpacity
-          style={{
-            marginLeft: 4,
-            borderRadius: 100,
-            width: 168,
-            height: 52,
-            justifyContent: 'center',
-            backgroundColor: '#193340'
-          }}>
-          <Text
-            style={{
-              fontSize: 17,
-              color: '#fff',
-              fontWeight: '500',
-              textAlign: 'center'
-            }}>
-            Next
-          </Text>
-        </TouchableOpacity>
-      </View>
-    </SafeAreaView>
-  )
-}
-
-export default SelectionScreen
