+++ src/screens/SucHab01Screens/PromptAudio.js
@@ -14,7 +14,6 @@ const PromptAudio = ({ route }) => {
     state: { responses }
   } = route.params
 
-  console.log(route.params, 'route.params audoe')
   const prompt = prompts[responses]
 
   const Audio = () => {
